pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
socat = ["alembic/*", "alembic.ini", "alembic/versions/*"]

[tool.setuptools.packages.find]
exclude=["*alembic*"]

[project]
name = "socat"
version = "0.0.1a1"
requires-python = ">=3.11"
dependencies = [
    "sqlmodel",
    "fastapi",
    "sqlalchemy[asyncio]",
    "uvicorn",
    "aiosqlite",
    "alembic",
    "pydantic_settings",
    "asyncer",
    "astropy",
    "scipy",
    "astroquery",
    "astropydantic"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff==0.14.5", # If you change this, also change the version in the CI
    "testcontainers",
    "coverage",
    "pytest-cov",
    "pytest-asyncio",
    "pre-commit",
    "httpx",
    "gevent"
]

[project.scripts]
socat-act-fits = "socat.ingest.actfits:main"
socat-websky-csv = "socat.ingest.webskycsv:main"
socat-migrate = "socat.migrate:migrate"

[tool.ruff.lint]
preview = true
extend-select = ["I", "DOC501"]
# ignore = ["D417", "D400", "D200", "D103", "D102", "D101", "D100"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = [
    "socat"
]
concurrency = ["gevent", "thread"]
